+++ src/libs/shouldEnableContextMenuEnterShortcut/index.native.ts
