+++ src/libs/shouldEnableContextMenuEnterShortcut/index.ts
